refactor: export words moved to controller
